requirements.txt
absl-py==2.0.0
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
dataclasses-json==0.6.2
decorator==4.4.2
distrax==0.1.3
dm-control==1.0.16
dm-tree==0.1.8
dmcgym @ git+https://github.com/WEIRDLabUW/dmcgym
docker-pycreds==0.4.0
etils==1.5.2
flax==0.7.2
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
glfw==2.6.2
gym==0.23.1
gym-notices==0.0.8
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
jax==0.4.13
jaxlib==0.4.13
markdown-it-py==3.0.0
marshmallow==3.20.1
matplotlib==3.7.4
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.1
moviepy==1.0.3
msgpack==1.0.7
mujoco==3.1.1
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.22.4
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.2.3
orbax-export==0.0.3
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
protobuf==4.25.0
psutil==5.9.6
Pygments==2.16.1
PyOpenGL==3.1.7
pytz
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
scipy==1.11.3
scikit-image==0.21.0
scikit-learn==1.3.2
sentry-sdk==1.35.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tensorflow==2.13.0
tensorflow-probability==0.21.0
tensorstore==0.1.49
toolz==0.12.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==2.1.0
wandb==0.16.0
zipp==3.17.0